dropvault-website/backend/requirements.txt
fastapi
uvicorn
sentence-transformers
numpy
pdfplumber
beautifulsoup4
lxml
python-multipart
playwright
pytesseract
pillow
opencv-python
cryptography
python-dotenv
